% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tttr}
\alias{tttr}
\title{Calculate TTTR Metric Score}
\usage{
tttr(input_file, monthly = FALSE)
}
\arguments{
\item{input_file}{Path to file containing travel time readings}

\item{monthly}{Calculate scores by month?}
}
\value{
A data.table of TTTR scores by TMC
}
\description{
A convenience wrapper for score(metric = "TTTR")
}
\examples{
\dontrun{
tttr("data/All_Vehicles/Readings.csv", monthly = TRUE)
}

}
