% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hpms}
\alias{hpms}
\title{Generate an HPMS Submission File}
\usage{
hpms(shp, lottr, tttr, year = NA)
}
\arguments{
\item{shp}{The sf shapefile containing the corresponding shapefile}

\item{lottr}{A data.table of LOTTR scores produced using \code{score(..., metric == "LOTTR")}}

\item{tttr}{A data.table of TTTR scores produced using \code{score(..., metric == "TTTR")}}

\item{year}{Year for which the file is being generated. Defaults to last year}
}
\description{
Generate an HPMS submission file in accordance with
https://www.fhwa.dot.gov/tpm/guidance/pm3_hpms.pdf
Requires the scores from score() to be run with verbose = TRUE
Writes the resulting file to out/hpms_year.txt
}
\examples{
\dontrun{
shp <- st_read("shp/Wyoming_2019/Wyoming.shp", stringsAsFactors = FALSE)
lottr <- score("data/All_Vehicles/al_tt_seconds.csv", metric = "LOTTR", verbose = TRUE)
tttr <- score("data/Trucks/aldot_2019_trucks.csv", metric = "TTTR", verbose = TRUE)
hpms(shp, lottr, tttr)
}

}
