% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lottr}
\alias{lottr}
\title{Calculate LOTTR Metric Score}
\usage{
lottr(input_file, monthly = FALSE)
}
\arguments{
\item{input_file}{Path to file containing travel time readings}

\item{monthly}{Calculate scores by month?}
}
\value{
A data.table of LOTTR scores by TMC
}
\description{
A convenience wrapper for score(metric = "LOTTR")
}
\examples{
\dontrun{
lottr("data/All_Vehicles/Readings.csv", monthly = TRUE)
}

}
