% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{score}
\alias{score}
\title{Calculate LOTTR or TTTR Metric Score}
\usage{
score(
  input_file = NULL,
  DT = NULL,
  metric = "LOTTR",
  period = "none",
  verbose = FALSE
)
}
\arguments{
\item{input_file}{Path to file containing travel time readings}

\item{DT}{A data.table of travel time readings (if input_file not supplied)}

\item{metric}{"LOTTR" or "TTTR"}

\item{period}{"none", "monthly", or "annual" specifies if the results should be
aggregated by time period. If "none" scores will be computed by TMC for all
data in the input file. "annual" only used if an input readings file contains
data for multiple years (discouraged due to annual shapefile updates)}

\item{verbose}{Provide diagnostic output and return all calculated values}
}
\value{
A data.table of LOTTR/TTTR scores by TMC
}
\description{
Calculate LOTTR / TTTR given a RITIS NPMRDS export of travel time data.
Data can be passed in as the path to the csv containing the travel time
readings or as a data.table of travel time readings.
input file must have header and format: 
tmc_code,measurement_tstamp,travel_time_seconds
 e.g. 
 > tmc_code,measurement_tstamp,travel_time_seconds
 > 116-04379,2019-01-01 00:00:00,44.78
 > 116-04379,2019-01-01 00:15:00,46.69
}
\examples{
\dontrun{
score("data/All_Vehicles/Readings.csv", metric = "LOTTR")
score("data/Trucks/Readings.csv", metric = "TTTR", period = "monthly")
}

}
